Move lsif types file.
